Refactor from binary into lib with examples
