+++ examples/server.rs
@@ -1,8 +1,4 @@
-extern crate ascii;
-extern crate futures;
-
 use anyhow::anyhow;
-use ascii::AsciiStr;
 use async_std::fs::File;
 use async_std::future::timeout;
 use async_std::net::Ipv4Addr;
@@ -13,9 +9,9 @@ use async_std::prelude::*;
 use futures::executor::block_on;
 use futures::executor::ThreadPool;
 use std::convert::TryFrom;
-use std::convert::TryInto;
 use std::string;
 use std::time::{Duration, Instant};
+use tftprust::packets::{TFTPMode, TFTPPacket};
 
 const MAX_DATA_SIZE: usize = 512;
 const DATA_HEADER_SIZE: usize = 4;
@@ -25,155 +21,9 @@ const MAX_ACK_TIMEOUT_SECS: u64 = 10;
 
 type TFTPBuffer = [u8; MAX_DATA_PACKET_SIZE];
 
-#[derive(PartialEq)]
-enum TFTPMode {
-    Ascii,
-    Octet,
-}
-
-enum TFTPPacket {
-    ReadRequest(string::String, TFTPMode),
-    WriteRequest(string::String, TFTPMode),
-    Data(u16, Vec<u8>),
-    Ack(u16),
-    Error(u16, string::String),
-}
-
-impl TryFrom<(TFTPBuffer, usize)> for TFTPPacket {
-    type Error = anyhow::Error;
-    fn try_from(value: (TFTPBuffer, usize)) -> anyhow::Result<Self> {
-        let data = value.0;
-        let length = value.1;
-
-        if length < 4 {
-            return Err(anyhow!("Not enough data for opcode + smallest packet"));
-        }
-
-        let opcode_bytes = data[..2].try_into();
-        if opcode_bytes.is_err() {
-            return Err(anyhow!("Failed to convert slice to array somehow?"));
-        }
-        let opcode = u16::from_be_bytes(opcode_bytes.unwrap());
-
-        match opcode {
-            // Read Request
-            1 => {
-                // There should only be two null bytes total
-                {
-                    let num_nulls = data[2..length].iter().filter(|&b| *b == 0).count();
-                    if num_nulls != 2 {
-                        return Err(anyhow!("Unexpected number of nulls"));
-                    }
-                }
-
-                // Second null byte should be the last
-                let last_null_index = length - 1;
-                if data[last_null_index] != 0 {
-                    return Err(anyhow!("Last byte of RRQ must be zero"));
-                }
-
-                let iter = data[..length].iter();
-
-                // Find first null byte
-                let first_null_index = iter.skip(2).position(|&b| b == 0);
-                if first_null_index.is_none() {
-                    return Err(anyhow!("Invalid RRQ - no first null byte"));
-                }
-                let first_null_index = first_null_index.unwrap() + 2;
-
-                // Grab the filename
-                let filename = AsciiStr::from_ascii(&data[2..first_null_index]);
-                if filename.is_err() {
-                    return Err(anyhow!("Invalid filename"));
-                }
-                let filename = filename.unwrap().trim();
-                if filename.is_empty() {
-                    return Err(anyhow!("Filename must not be empty"));
-                }
-                // TODO validate filename is legal filename characters?
-
-                // Grab the mode string
-                let mode = AsciiStr::from_ascii(&data[first_null_index + 1..last_null_index]);
-                if mode.is_err() {
-                    return Err(anyhow!("Invalid mode string"));
-                }
-                let mode = mode.unwrap().to_ascii_lowercase();
-
-                match mode.as_ref() {
-                    "octet" => {
-                        return Ok(TFTPPacket::ReadRequest(
-                            filename.to_string(),
-                            TFTPMode::Octet,
-                        ));
-                    }
-                    "netascii" => {
-                        return Ok(TFTPPacket::ReadRequest(
-                            filename.to_string(),
-                            TFTPMode::Ascii,
-                        ));
-                    }
-                    _ => {
-                        return Err(anyhow!("Invalid/unsupported mode"));
-                    }
-                }
-            }
-            // Write Request
-            2 => {}
-            // Data
-            3 => {}
-            // Ack
-            4 => {
-                // There should only be four bytes
-                if length != 4 {
-                    return Err(anyhow!("ACK packets must be exactly four bytes"));
-                }
-
-                // Read the block number
-                let block_num_array: [u8; 2] = [data[2], data[3]];
-                let block_num = u16::from_be_bytes(block_num_array);
-                return Ok(TFTPPacket::Ack(block_num));
-            }
-            // Error
-            5 => {}
-            // WTF
-            _ => {
-                return Err(anyhow!("Unknown opcode"));
-            }
-        }
-
-        Ok(TFTPPacket::ReadRequest(
-            String::from("asdf"),
-            TFTPMode::Octet,
-        ))
-    }
-}
-
 pub struct TFTPServer {
     socket: UdpSocket,
     buf: TFTPBuffer,
-}
-
-impl std::fmt::Debug for TFTPPacket {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TFTPPacket::ReadRequest(filename, mode) => {
-                write!(f, "ReadRequest('{}', {:?})", filename, mode)
-            }
-            TFTPPacket::WriteRequest(_, _) => write!(f, "WriteRequest"),
-            TFTPPacket::Data(_, _) => write!(f, "Data"),
-            TFTPPacket::Ack(_) => write!(f, "Ack"),
-            TFTPPacket::Error(_, _) => write!(f, "Error"),
-        }
-    }
-}
-
-impl std::fmt::Debug for TFTPMode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TFTPMode::Ascii => write!(f, "NetASCII"),
-            TFTPMode::Octet => write!(f, "octet"),
-        }
-    }
 }
 
 impl TFTPServer {
@@ -193,10 +43,11 @@ impl TFTPServer {
     //
     fn main_loop_iter(&mut self, pool: &ThreadPool) -> anyhow::Result<()> {
         let (num_recv_bytes, sender) = block_on(self.socket.recv_from(&mut self.buf))?;
+        let data = self.buf[0..num_recv_bytes].to_vec();
         // println!("Received {} bytes from {:?}: {:?}", num_recv_bytes, sender, &self.buf[..num_recv_bytes]);
 
         // Parse bytes into packet
-        let packet = TFTPPacket::try_from((self.buf, num_recv_bytes))?;
+        let packet = TFTPPacket::try_from(data)?;
 
         eprintln!("Received {:?} from {:?}", packet, sender);
 
@@ -299,7 +150,9 @@ impl TFTPServer {
             let recv_result = r.unwrap();
             let num_read_socket = recv_result?;
 
-            let packet = TFTPPacket::try_from((to_recv, num_read_socket));
+            let data = to_recv[0..num_read_socket].to_vec();
+
+            let packet = TFTPPacket::try_from(data);
             if packet.is_err() {
                 eprintln!(
                     "Received {} unparsable bytes: {:?}. Error: {}",
@@ -335,5 +188,15 @@ pub fn make_tftp_server<A: ToSocketAddrs>(addr: A) -> TFTPServer {
     TFTPServer {
         socket: block_on(UdpSocket::bind(addr)).unwrap(),
         buf: [0; MAX_DATA_PACKET_SIZE],
+    }
+}
+
+fn main() {
+    let mut server = make_tftp_server("0.0.0.0:69");
+
+    let res = server.run();
+    if res.is_err() {
+        eprintln!("Server stopped with error: {:?}", res.err());
+        std::process::exit(1);
     }
 }
