+++ src/lib.rs
@@ -1,0 +1,2 @@
+/// Structs and functions for representing TFTP packets and serializing/deserializing them.
+pub mod packets;
